Refactor kernel utilities layout
