+++ kernel/Makefile
@@ -31,6 +31,7 @@ KERNEL_ELF = $(BUILD_DIR)/kernel/exos.elf
 
 SRC_C  = $(wildcard source/*.c) \
          $(wildcard source/network/*.c) \
+         $(wildcard source/utils/*.c) \
          $(wildcard source/test/*.c) \
          $(wildcard source/autotest/*.c) \
          ../third/bcrypt/blowfish.c \
