+++ kernel/include/Kernel.h
@@ -28,7 +28,7 @@
 /***************************************************************************/
 
 #include "Base.h"
-#include "Database.h"
+#include "utils/Database.h"
 #include "FileSystem.h"
 #include "Heap.h"
 #include "I386.h"
