+++ kernel/include/Path.h
@@ -28,7 +28,7 @@
 /***************************************************************************/
 
 #include "FileSystem.h"
-#include "StringArray.h"
+#include "utils/StringArray.h"
 
 /***************************************************************************/
