+++ kernel/include/Socket.h
@@ -30,7 +30,7 @@
 #include "Network.h"
 #include "network/TCP.h"
 #include "ID.h"
-#include "CircularBuffer.h"
+#include "utils/CircularBuffer.h"
 
 /************************************************************************/
 // Socket Address Family
