+++ kernel/include/utils/StringArray.h
