+++ kernel/source/Shell.c
@@ -25,7 +25,7 @@
 #include "../include/ACPI.h"
 #include "../include/Base.h"
 #include "../include/Clock.h"
-#include "../include/CommandLineEditor.h"
+#include "../include/utils/CommandLineEditor.h"
 #include "../include/Console.h"
 #include "../include/Disk.h"
 #include "../include/Endianness.h"
@@ -45,7 +45,7 @@
 #include "../include/Script.h"
 #include "../include/StackTrace.h"
 #include "../include/String.h"
-#include "../include/StringArray.h"
+#include "../include/utils/StringArray.h"
 #include "../include/System.h"
 #include "../include/User.h"
 #include "../include/UserAccount.h"
