+++ kernel/source/UserAccount.c
@@ -25,8 +25,8 @@
 #include "../include/UserAccount.h"
 
 #include "../include/Clock.h"
-#include "../include/Crypt.h"
-#include "../include/Database.h"
+#include "../include/utils/Crypt.h"
+#include "../include/utils/Database.h"
 #include "../include/Heap.h"
 #include "../include/Helpers.h"
 #include "../include/Kernel.h"
