+++ kernel/source/utils/CRC64.c
@@ -22,8 +22,8 @@
 
 \************************************************************************/
 
-#include "../include/Crypt.h"
-#include "../include/String.h"
+#include "utils/Crypt.h"
+#include "String.h"
 
 /************************************************************************/
