+++ kernel/source/utils/StateMachine.c
@@ -22,7 +22,7 @@
 
 \************************************************************************/
 
-#include "../include/StateMachine.h"
+#include "utils/StateMachine.h"
 
 /************************************************************************/
